Finish FileInput
