+++ compass/src/components/Shared/FileInput/FileInput.js
@@ -1,25 +1,23 @@
 import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
-import { FormMessage } from 'fundamental-react';
 import './FileInput.scss';
 
 FileInput.propTypes = {
   fileInputChanged: PropTypes.func.isRequired,
-  file: PropTypes.object,
-  error: PropTypes.string,
   availableFormatsMessage: PropTypes.string,
   acceptedFileFormats: PropTypes.string.isRequired,
+  required: PropTypes.bool,
 };
 
 export default function FileInput({
   fileInputChanged,
-  file,
-  error,
   availableFormatsMessage,
   acceptedFileFormats,
   inputRef,
+  required,
 }) {
+  const [fileName, setFileName] = useState('');
   const [draggingOverCounter, setDraggingCounter] = useState(0);
 
   // needed for onDrag to fire
@@ -28,11 +26,16 @@ export default function FileInput({
     e.preventDefault();
   }
 
+  function fileChanged(file) {
+    setFileName(file ? file.name : '');
+    fileInputChanged(file);
+  }
+
   function drop(e) {
     setDraggingCounter(0);
     e.preventDefault();
     e.nativeEvent.stopImmediatePropagation(); // to avoid event.js error
-    fileInputChanged(e.dataTransfer.files[0]);
+    fileChanged(e.dataTransfer.files[0]);
   }
 
   const labelClass = classNames('fd-asset-upload__label', {
@@ -47,14 +50,15 @@ export default function FileInput({
       onDragLeave={() => setDraggingCounter(draggingOverCounter - 1)}
       onDragOver={dragOver}
     >
-      {!!file && <p className="fd-asset-upload__file-name">{file.name}</p>}
+      {!!fileName && <p className="fd-asset-upload__file-name">{fileName}</p>}
       <input
         ref={inputRef}
         type="file"
         id="file-upload"
+        onChange={e => fileChanged(e.target.files[0])}
         className="fd-asset-upload__input"
-        onChange={e => fileInputChanged(e.target.files[0])}
         accept={acceptedFileFormats}
+        required={required}
       />
       <label htmlFor="file-upload" className={labelClass}>
         <span className="fd-asset-upload__text">Browse</span>
@@ -63,7 +67,6 @@ export default function FileInput({
           <p className="fd-asset-upload__message">{availableFormatsMessage}</p>
         )}
       </label>
-      {!!error && <FormMessage type="error">{error} </FormMessage>}
     </div>
   );
 }
